// backend/src/db/schema.d.ts
export declare const participations: import("drizzle-orm/pg-core").PgTableWithColumns<{
    name: "participations";
    schema: undefined;
    columns: {
        id: import("drizzle-orm/pg-core").PgBuildColumn<"participations", import("drizzle-orm/pg-core").SetIsPrimaryKey<import("drizzle-orm/pg-core").SetHasDefault<import("drizzle-orm/pg-core").PgUUIDBuilder>>, {
            name: string;
            tableName: "participations";
            dataType: "string uuid";
            data: string;
            driverParam: string;
            notNull: true;
            hasDefault: true;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            identity: undefined;
            generated: undefined;
        }>;
        userId: import("drizzle-orm/pg-core").PgBuildColumn<"participations", import("drizzle-orm/pg-core").SetNotNull<import("drizzle-orm/pg-core").PgUUIDBuilder>, {
            name: string;
            tableName: "participations";
            dataType: "string uuid";
            data: string;
            driverParam: string;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            identity: undefined;
            generated: undefined;
        }>;
        activityId: import("drizzle-orm/pg-core").PgBuildColumn<"participations", import("drizzle-orm/pg-core").SetNotNull<import("drizzle-orm/pg-core").PgUUIDBuilder>, {
            name: string;
            tableName: "participations";
            dataType: "string uuid";
            data: string;
            driverParam: string;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            identity: undefined;
            generated: undefined;
        }>;
        status: import("drizzle-orm/pg-core").PgBuildColumn<"participations", import("drizzle-orm/pg-core").SetHasDefault<import("drizzle-orm/pg-core").SetNotNull<import("drizzle-orm/pg-core").PgEnumColumnBuilder<["PENDING", "ACCEPTED", "REFUSED"]>>>, {
            name: string;
            tableName: "participations";
            dataType: "string enum";
            data: "PENDING" | "ACCEPTED" | "REFUSED";
            driverParam: string;
            notNull: true;
            hasDefault: true;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: ["PENDING", "ACCEPTED", "REFUSED"];
            identity: undefined;
            generated: undefined;
        }>;
        createdAt: import("drizzle-orm/pg-core").PgBuildColumn<"participations", import("drizzle-orm/pg-core").SetNotNull<import("drizzle-orm/pg-core").SetHasDefault<import("drizzle-orm/pg-core").PgTimestampBuilder>>, {
            name: string;
            tableName: "participations";
            dataType: "object date";
            data: Date;
            driverParam: string;
            notNull: true;
            hasDefault: true;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            identity: undefined;
            generated: undefined;
        }>;
    };
    dialect: "pg";
}>;